Move some functionality from VdevFile to VdevBlock
Move the Manager, Label, and Status structs.  They don't have any
file-specific code, so it makes sense to put them here so we can support
multiple leaf types later.
