+++ bfffs-core/src/mirror.rs
@@ -531,6 +531,10 @@ impl Mirror {
         }
     }
 
+    pub fn uuid(&self) -> Uuid {
+        self.uuid
+    }
+
     pub fn write_at(&self, buf: IoVec, lba: LbaT) -> BoxVdevFut
     {
         let fut = self.children.iter().filter_map(|blockdev| {
@@ -601,10 +605,6 @@ impl Vdev for Mirror {
         .try_collect::<Vec<_>>()
         .map_ok(drop);
         Box::pin(fut)
-    }
-
-    fn uuid(&self) -> Uuid {
-        self.uuid
     }
 
     fn zone_limits(&self, zone: ZoneT) -> (LbaT, LbaT) {
@@ -752,8 +752,9 @@ mock! {
         pub fn read_long(&self, len: LbaT, lba: LbaT)
             -> Pin<Box<dyn Future<Output=Result<Box<dyn Iterator<Item=DivBufShared> + Send>>> + Send>>;
         pub fn read_spacemap(&self, buf: IoVecMut, idx: u32) -> BoxVdevFut;
-        pub fn status(&self) -> Status;
         pub fn readv_at(&self, bufs: SGListMut, lba: LbaT) -> BoxVdevFut;
+        pub fn status(&self) -> Status;
+        pub fn uuid(&self) -> Uuid;
         pub fn write_at(&self, buf: IoVec, lba: LbaT) -> BoxVdevFut;
         pub fn write_label(&self, labeller: LabelWriter) -> BoxVdevFut;
         pub fn write_spacemap(&self, sglist: SGList, idx: u32, block: LbaT)
@@ -765,7 +766,6 @@ mock! {
         fn optimum_queue_depth(&self) -> u32;
         fn size(&self) -> LbaT;
         fn sync_all(&self) -> BoxVdevFut;
-        fn uuid(&self) -> Uuid;
         fn zone_limits(&self, zone: ZoneT) -> (LbaT, LbaT);
         fn zones(&self) -> ZoneT;
     }
