+++ bfffs-core/tests/functional/vdev_file.rs
@@ -138,11 +138,6 @@ mod basic {
     }
 
     #[rstest]
-    fn path(harness: Harness) {
-        assert_eq!(harness.vdev.path(), harness.path.as_path());
-    }
-
-    #[rstest]
     fn size(harness: Harness) {
         assert_eq!(harness.vdev.size(), 16_384);
     }
@@ -361,179 +356,5 @@ mod dev {
 
         // Must drop vdev before md
         drop(vd);
-    }
-}
-
-mod persistence {
-    use bfffs_core::{
-        *,
-        label::*,
-        vdev::*,
-        vdev_file::*
-    };
-    use pretty_assertions::assert_eq;
-    use rstest::{fixture, rstest};
-    use std::{
-        fs,
-        io::Read,
-        num::NonZeroU64,
-        os::unix::fs::FileExt,
-        path::PathBuf
-    };
-    use tempfile::{Builder, TempDir};
-
-    const GOLDEN: [u8; 72] = [
-        // First 16 bytes are file magic
-        0x42, 0x46, 0x46, 0x46, 0x53, 0x20, 0x56, 0x64, // BFFFS Vd
-        0x65, 0x76, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // ev......
-        // Next 8 bytes are a checksum
-        0x2e, 0x43, 0xc2, 0x5d, 0x1f, 0x55, 0x20, 0x3b,
-        // Next 8 bytes are the contents length, in BE
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28,
-        // The rest is a serialized VdevFile::Label object.
-        // First comes the VdevFile's UUID.
-        0x3f, 0xa1, 0xf6, 0xb9, 0x54, 0xb1, 0x4a, 0x10,
-        0xbc, 0x6b, 0x5b, 0x2a, 0x15, 0xe8, 0xa0, 0x3d,
-        // Then the number of LBAS per zone
-        0xbe, 0xba, 0x7e, 0x1a, 0xef, 0xbe, 0xad, 0xde,
-        // Then the number of LBAs as a 64-bit number
-        0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        // Finally the number of LBAs reserved for the spacemap
-        0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    ];
-
-    struct Harness {
-        path: PathBuf,
-        _tempdir: TempDir
-    }
-
-    #[fixture]
-    fn harness() -> Harness {
-        let len = 1 << 26;  // 64MB
-        let tempdir = Builder::new()
-            .prefix("test_vdev_file_persistence")
-            .tempdir()
-            .unwrap();
-        let filename = tempdir.path().join("vdev");
-        let file = fs::File::create(&filename).unwrap();
-        file.set_len(len).unwrap();
-        Harness{path: filename, _tempdir: tempdir}
-    }
-
-    /// Open the golden master label
-    #[rstest]
-    #[tokio::test]
-    async fn open(harness: Harness) {
-        let golden_uuid = Uuid::parse_str(
-            "3fa1f6b9-54b1-4a10-bc6b-5b2a15e8a03d").unwrap();
-        {
-            let f = std::fs::OpenOptions::new()
-                .write(true)
-                .open(harness.path.clone()).unwrap();
-            let offset0 = 0;
-            f.write_all_at(&GOLDEN, offset0).unwrap();
-            let offset1 = 4 * BYTES_PER_LBA as u64;
-            f.write_all_at(&GOLDEN, offset1).unwrap();
-        }
-        let (vdev, _label_reader) = VdevFile::open(harness.path).await.unwrap();
-        assert_eq!(vdev.size(), 16_384);
-        assert_eq!(vdev.uuid(), golden_uuid);
-    }
-
-    // Open a device with only corrupted labels
-    #[rstest]
-    #[tokio::test]
-    async fn open_ecksum(harness: Harness) {
-        {
-            let f = std::fs::OpenOptions::new()
-                .write(true)
-                .open(harness.path.clone()).unwrap();
-            let offset0 = 0;
-            f.write_all_at(&GOLDEN, offset0).unwrap();
-            let offset1 = 4 * BYTES_PER_LBA as u64;
-            f.write_all_at(&GOLDEN, offset1).unwrap();
-            let zeros = [0u8; 1];
-            // Corrupt the labels' checksum
-            f.write_all_at(&zeros, offset0 + 16).unwrap();
-            f.write_all_at(&zeros, offset1 + 16).unwrap();
-        }
-        let e = VdevFile::open(harness.path).await
-            .err()
-            .expect("Opening the file should've failed");
-        assert_eq!(e, Error::EINTEGRITY);
-    }
-
-    /// Open a device that only has one valid label, the first one
-    #[rstest]
-    #[tokio::test]
-    async fn open_first_label_only(harness: Harness) {
-        let golden_uuid = Uuid::parse_str(
-            "3fa1f6b9-54b1-4a10-bc6b-5b2a15e8a03d").unwrap();
-        {
-            let f = std::fs::OpenOptions::new()
-                .write(true)
-                .open(harness.path.clone()).unwrap();
-            let offset0 = 0;
-            f.write_all_at(&GOLDEN, offset0).unwrap();
-        }
-        let (vdev, _label_reader) = VdevFile::open(harness.path).await.unwrap();
-        assert_eq!(vdev.size(), 16_384);
-        assert_eq!(vdev.uuid(), golden_uuid);
-    }
-
-    // Open a device without a valid label
-    #[rstest]
-    #[tokio::test]
-    async fn open_invalid(harness: Harness) {
-        let e = VdevFile::open(harness.path).await
-            .err()
-            .expect("Opening the file should've failed");
-        assert_eq!(e, Error::EINVAL);
-    }
-
-    /// Open a device that only has one valid label, the second one
-    #[rstest]
-    #[tokio::test]
-    async fn open_second_label_only(harness: Harness) {
-        let golden_uuid = Uuid::parse_str(
-            "3fa1f6b9-54b1-4a10-bc6b-5b2a15e8a03d").unwrap();
-        {
-            let f = std::fs::OpenOptions::new()
-                .write(true)
-                .open(harness.path.clone()).unwrap();
-            let offset1 = 4 * BYTES_PER_LBA as u64;
-            f.write_all_at(&GOLDEN, offset1).unwrap();
-        }
-        let (vdev, _label_reader) = VdevFile::open(harness.path).await.unwrap();
-        assert_eq!(vdev.size(), 16_384);
-        assert_eq!(vdev.uuid(), golden_uuid);
-    }
-
-    // Write the label, and compare to a golden master
-    #[rstest]
-    #[tokio::test]
-    async fn write_label(harness: Harness) {
-        let lbas_per_zone = NonZeroU64::new(0xdead_beef_1a7e_babe);
-        let vdev = VdevFile::create(harness.path.clone(), lbas_per_zone)
-            .unwrap();
-        let label_writer = LabelWriter::new(0);
-        vdev.write_label(label_writer).await.unwrap();
-
-        let mut f = std::fs::File::open(harness.path).unwrap();
-        let mut v = vec![0; 4096];
-        f.read_exact(&mut v).unwrap();
-        // Uncomment this block to save the binary label for inspection
-        /* {
-            use std::fs::File;
-            use std::io::Write;
-            let mut df = File::create("/tmp/label.bin").unwrap();
-            df.write_all(&v[..]).unwrap();
-            println!("UUID is {}", vdev.uuid());
-        } */
-        // Compare against the golden master, skipping the checksum and UUID
-        // fields
-        assert_eq!(&v[0..16], &GOLDEN[0..16]);
-        assert_eq!(&v[24..32], &GOLDEN[24..32]);
-        assert_eq!(&v[48..GOLDEN.len()], &GOLDEN[48..GOLDEN.len()]);
     }
 }
